refactor: add sharedcomponents namespace
