+++ src/client/components/player-client.ts
@@ -2,7 +2,7 @@ import { Component, Components } from "@flamework/components";
 
 import { AbstractPlayer } from "shared/components/abstract-player";
 import { OnLocalCharacterAdded } from "../../../types/lifecycles";
-import { CharacterClient } from "./character-client";
+import { Character } from "./character";
 
 @Component({
 	tag: "Player",
@@ -11,22 +11,22 @@ export class PlayerClient
 	extends AbstractPlayer
 	implements OnLocalCharacterAdded
 {
-	private character?: CharacterClient;
+	private character?: Character;
 
 	public constructor(private components: Components) {
 		super();
 	}
 
 	public onLocalCharacterAdded(model: Model): void {
-		this.character = this.components.addComponent<CharacterClient>(model);
+		this.character = this.components.addComponent<Character>(model);
 	}
 
 	public onLocalCharacterRemoving(model: Model): void {
 		this.character = undefined;
-		this.components.removeComponent<CharacterClient>(model);
+		this.components.removeComponent<Character>(model);
 	}
 
-	public getCharacter(): CharacterClient {
+	public getCharacter(): Character {
 		if (this.character === undefined) error("character unavailable");
 		return this.character;
 	}
