+++ src/server/services/hit-service.ts
@@ -1,6 +1,6 @@
 import { Components } from "@flamework/components";
 import { OnStart, Service } from "@flamework/core";
-import { CharacterServer } from "server/components/character/character-server";
+import { Character } from "server/components/character/character";
 import { Events } from "server/networking";
 
 @Service()
@@ -17,18 +17,18 @@ export class HitService implements OnStart {
 
 	private registerHit(player: Player, character: Model): void {
 		if (player.Character === undefined) return;
-		const hitter = this.components.getComponent<CharacterServer>(
+		const hitter = this.components.getComponent<Character>(
 			player.Character,
 		);
 		if (hitter === undefined) return;
-		const victim = this.components.getComponent<CharacterServer>(character);
+		const victim = this.components.getComponent<Character>(character);
 		if (victim === undefined) return;
 		if (!this.canHit(hitter, victim)) return;
 
 		victim.takeDamage(10);
 	}
 
-	private canHit(hitter: CharacterServer, victim: CharacterServer): boolean {
+	private canHit(hitter: Character, victim: Character): boolean {
 		return (
 			hitter.attributes.isAlive &&
 			!hitter.attributes.isKnocked &&
