+++ src/server/services/reset-service.ts
@@ -1,6 +1,6 @@
 import { Components } from "@flamework/components";
 import { Dependency, OnStart, Service } from "@flamework/core";
-import { CharacterServer } from "server/components/character/character-server";
+import { Character } from "server/components/character/character";
 import { Events } from "server/networking";
 
 @Service()
@@ -11,7 +11,7 @@ export class ResetService implements OnStart {
 		Events.reset.connect((player) => {
 			if (!player.Character) return;
 
-			const character = this.components.getComponent<CharacterServer>(
+			const character = this.components.getComponent<Character>(
 				player.Character,
 			);
 			if (!character) return;
